Some bugs fixed and Add font and make a stability better

+++ CardGame/Vector.h
@@ -4,9 +4,9 @@
 class Vector {
 private:
 	void _Forward();
-	int dx, dy;
+	float dx, dy;
 public:
-	int x, y, speed;
+	float x, y, speed;
 	Vector(int _x, int _y) : x(_x), y(_y), dx(-1), dy(-1), speed(-1) {}
 	Vector() : x(-1), y(-1), dx(-1), dy(-1), speed(-1) {}
